refactor: extract discussions api calls into api/client
