+++ package.json
@@ -138,6 +138,7 @@
     "css-loader": "7.1.1",
     "electron": "30.0.1",
     "electron-builder": "24.13.3",
+    "graphql-tag": "2.12.6",
     "husky": "9.0.11",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
